On plug, launch a TX/RX ring-polling thread
This allows TX to work immediately, before the application is
ready to call listen(). All RXed packets are sent to the
receive_callback which is initially (fun buffer -> return ())
i.e. the packets are dropped. Previously the packets would be queued
in the RX ring and then dropped by the backend once the ring
filled up.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>