Ensure that we poll and refill the RX ring in the right order.
This avoids a deadlock due to the old code being able to refill
the ring, draining it entirely (in rx_poll) and then blocking in
Activations.  There is a larger fix required to address that we
are dropping into Activations too much in this code, since we are
not checking if work is coming in on the ring after a poll.

Bug found by Masoud Koleini in #15