Wait for the backend to enter the connected state
Previously we jumped straight into the connected state, read the backend
features and started transmitting packets and events. Some backends
(including at least the Linux netback) are not prepared to receive anything
before they have also transitioned into the connected state.

I believe this fixes #20

There is a chance that this doesn't actually fix the issue but simply masks
it with additional delay. However: (i) the delay seems very short; and
(ii) this is the right thing to do anyway.

Signed-off-by: David Scott <dave.scott@citrix.com>