Import some modules from djs55/refactor-modules
This brings in Flag, RX, TX and ResultM (which now uses the new "result"
type instead of a custom one).

Now we only need to disable unused value warnings in the protocol files.
Removed all unused code that this discovered.

The public API is unchanged and the behaviour should be unchanged,
except that it may be more fussy about invalid flags and offsets.