Generate a unique ID for each Netfront request
Previously we used the grant ref, but this isn't unique. This could
possibly cause pages to be returned to the free pool while they were
still being transmitted.

Since IDs must be 16-bit, this patch also limits the size of the pool
(to 256 pages, which is roughly the size of the ring).