Don't assume the first interface is meant when id can't be decoded.
Previously, `connect` would assume that the first available interface
was the correct one to use if the `id` parameter couldn't be matched.
This could lead to information leakage or other unexpected (and
difficult to debug) behaviors.  Instead, don't assume this and return an
error if the id doesn't make sense.

Signed-off-by: Mindy Preston <mindy.preston@docker.com>