Log exceptions from the listen callback
Since version 1.5 we've been passing exceptions to the user's Lwt
exception handler rather than printing to the console. It seems that
several users decided to stay with 1.4 rather than configure a handler.
This patch uses the new logging support to log an error and continue.
In addition, it also dumps out the frame that caused the crash.

Closes #39