Support client-initiated shutdown
Commit cc77a1a829bf3 noted "Not sure whether we should be watching
the frontend or backend state, but this works with QubesOS at least."

It seems that if we want to allow Qubes guests to shut down their own
interfaces then we need to be watching the frontend instead (or maybe
as well).

This is important for HVM guests, as the stub domain tries to shut down
the interface while booting. For example, if you try to install Debian 9
it will hang after selecting the install option from the boot menu.
The guest's stub-domain log shows e.g.

    close network: backend at /local/domain/2/backend/vif/11/0

as the last output in this case.