Netback should not initialise the shared rings
Netfront is responsible for this ("note that only the front initializes
the shared ring", from `xen/include/public/io/ring.h`).
If netback does it too, it will overwrite the client's data.

In particular, this prevented Qubes disposable Linux VMs from working
when restored. They would populate the ring with RX pages before handing
it to the Mirage netback, which would then reinitialise the ring and
lose them all. It would then wait forever for Linux to provide it with
some pages to write to.